+++ oio_rest/oio_rest/db/sql/declarations/templates/as_search_dokument_mixin.jinja.sql
@@ -1,3 +1,6 @@
+-- SPDX-FileCopyrightText: 2015-2020 Magenta ApS
+-- SPDX-License-Identifier: MPL-2.0
+
 --/**********************************************************//
 --Filtration on variants and document parts (dele)
 --/**********************************************************//