+++ oio_rest/oio_rest/db/sql/invocations/templates/create_object.sql
@@ -1,3 +1,5 @@
+-- SPDX-FileCopyrightText: 2018-2020 Magenta ApS
+-- SPDX-License-Identifier: MPL-2.0
 
 SELECT * from as_create_or_import_{{ class_name|lower }}(
     {{registration}},