+++ oio_rest/tests/sql/test_remove_nulls_in_array_klasse.sql
@@ -1,9 +1,5 @@
--- Copyright (C) 2015 Magenta ApS, https://magenta.dk.
--- Contact: info@magenta.dk.
---
--- This Source Code Form is subject to the terms of the Mozilla Public
--- License, v. 2.0. If a copy of the MPL was not distributed with this
--- file, You can obtain one at http://mozilla.org/MPL/2.0/.
+-- SPDX-FileCopyrightText: 2015-2020 Magenta ApS
+-- SPDX-License-Identifier: MPL-2.0
 
 --SELECT * FROM runtests('test'::name);
 CREATE OR REPLACE FUNCTION test.test_remove_nulls_in_array_klasse()